feat: refactor documentation structure by consolidating docmeta functionality into doc-type and removing outdated code
